requirements.txt
fastapi[standard]>=0.111.0
motor>=3.4.0
openai>=1.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
uvicorn>=0.30.1
